pyproject.toml
[project]
name = "clofy-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "basedpyright>=1.29.2",
    "pre-commit>=3.3.3",
    "ruff>=0.11.11",
    "invoke>=2.2.0",
    "pytest>=8.0.0",
]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# --- Add/Modify this section for Setuptools/Packaging ---
[tool.setuptools]
# Explicitly list the packages to include.
# This prevents setuptools from trying to automatically discover packages
# which causes conflicts with directories like 'static'.
packages = ["app"] # Only include 'app' as a Python package.

